refactor: moved correlation analysis function for preprocessed datasets
